requirements.txt
google-generativeai
snscrape
aiofiles
